pyproject.toml
[project]
name = "spotter"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django==3.2.23",
    "djangorestframework==3.14.0",
    "openrouteservice==2.3.3",
    "pytest-django==4.5.2",
    "pytest==7.4.3",
    "requests==2.31.0",
    "pytest-mock==3.11.1",
    "vcrpy==4.3.1",
    "coverage==7.3.2",
    "python-dateutil==2.8.2",
    "pytz==2023.3.post1",
    "sqlparse==0.4.4",
    "typing-extensions==4.7.1",
    "asgiref==3.7.2",
    "numpy>=2.2.0",
    "pandas>=2.2.3",
    "geopy>=2.4.1",
    "googlemaps>=4.10.0",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.0.1",
    "folium>=0.19.2",
]

[project.optional-dependencies]
dev-dependencies = [
    "black>=24.10.0",
    "flake8>=7.1.1",
]
